Centralize flash type definitions
Moves shared `flash` TypeScript types into a new package-level `types.d.ts` and updates components to import from the centralized source instead of declaring local interfaces. `src/types.d.ts` and Chatbot types now re-export from the new file, preserving compatibility while consolidating definitions. Also adds `types.d.ts` to `tsconfig.json` includes and introduces `I_Response` with a deprecated `I_Resonse` alias for backward compatibility.
