+++ packages/flash/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@nx/nx-flash",
+  "name": "@nx/flash",
   "version": "3.0.0",
   "description": "NX Flash: Macromedia Flash-inspired stage and MovieClip component system",
   "private": true,
@@ -8,7 +8,10 @@
   "scripts": {
     "test": "../../apps/www/node_modules/.bin/tsx --test tests/**/*.test.ts",
     "type-check": "tsc --noEmit",
-    "lint": "echo 'No lint configured yet'"
+    "lint": "echo 'No lint configured yet'",
+    "example": "vite --config demo/vite.config.ts",
+    "example:build": "vite build --config demo/vite.config.ts",
+    "example:preview": "vite preview --config demo/vite.config.ts"
   },
   "peerDependencies": {
     "@emotion/react": ">=11.0.0",
@@ -23,10 +26,12 @@
     "redux": ">=4.0.0"
   },
   "devDependencies": {
+    "@vitejs/plugin-react": "^5.1.0",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "next": "*",
     "redux": "^5",
-    "typescript": "^5"
+    "typescript": "^5",
+    "vite": "^7.1.2"
   }
 }
