+++ packages/flash/src/Flash.tsx
@@ -1,10 +1,6 @@
 import React from 'react';
 import styles from './lib/Flash.module.css';
-
-export interface I_Flash {
-    id?: string;
-    children?: React.ReactNode;
-}
+import type { I_Flash } from './types';
 
 export const Flash: React.FC<I_Flash> = ({ children, id }) => {
     return (
