+++ packages/flash/src/MovieClips/Chatbot/types.d.ts
@@ -1,30 +1,1 @@
-import React from 'react';
-
-export interface I_Chunk {
-    [key: string]: any;
-};
-
-
-
-// Interface defining parameters and options for Chatbot
-export interface I_Chatbot {
-    // Define required parameters here
-    // Example: userId: string;
-    id?: string;
-    title?: string;
-    logo?: React.ReactNode;
-
-
-    // Optional options
-    // Example: theme?: string;
-    [key: string]: any; // Placeholder for extensibility
-}
-
-
-// Optional options
-export interface I_Resonse {
-    text: string;
-    from: 'user' | 'bot';
-    avatar?: React.ReactNode;
-    [key: string]: any; // Placeholder for extensibility
-}
+export type { I_Chatbot, I_Chunk, I_Response, I_Resonse } from '../../../types';
