+++ packages/flash/src/types.d.ts
@@ -1,48 +1,1 @@
-// Augment ImportMeta to support import.meta.hot (for Vite/webpack HMR)
-declare global {
-  interface ImportMeta {
-    hot?: {
-      accept: (cb: () => void) => void;
-    };
-  }
-}
-import React from 'react';
-
-export type TState = {
-  [key: string]: any;
-}
-
-export type TMovieClip = {
-  id?: string;
-  children?: React.ReactNode;
-  style?: any;
-  border?: boolean;
-  width?: number | string;
-  height?: number | string;
-  position?:
-  | 'top-left'
-  | 'top-middle'
-  | 'top-right'
-  | 'middle-left'
-  | 'middle-right'
-  | 'bottom-left'
-  | 'bottom-middle'
-  | 'bottom-right';
-};
-
-export type TFlashConfig = {
-  width?: number | string;
-  height?: number | string;
-  [key: string]: any;
-};
-
-export type TStage = {
-  id?: string;
-  width?: number | string;
-  height?: number | string;
-  children?: React.ReactNode;
-};
-
-export type TTrace = {
-  children?: React.ReactNode;
-};
+export * from '../types';
